requirements.txt
Flask~=2.2.3
pyppeteer
urllib3
requests~=2.28.2
flask-sqlalchemy
python-dotenv~=1.0.0